fix: replace toSql with AST2SQL
